# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MinePilot API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plugins:
    get:
      operationId: listPlugins
      tags: [plugins]
      summary: List available plugins in marketplace
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Plugin list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plugin"
  /plugins/installed:
    get:
      operationId: listInstalledPlugins
      tags: [plugins]
      summary: List installed plugins for a server
      security:
        - bearerAuth: []
      parameters:
        - name: serverId
          in: query
          required: false
          schema:
            type: integer
            nullable: true
      responses:
        "200":
          description: Installed plugins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstalledPlugin"
  /plugins/install:
    post:
      operationId: installPlugin
      tags: [plugins]
      summary: Install a plugin on a server
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallPluginRequest"
      responses:
        "200":
          description: Plugin installed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Plugin:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        version:
          type: string
        author:
          type: string
        category:
          type: string
        downloadUrl:
          type: string
        tags:
          type: array
          items:
            type: string
      required: [id, name, description, version, author, category, downloadUrl, tags]
    InstalledPlugin:
      type: object
      properties:
        id:
          type: integer
        serverId:
          type: integer
        pluginId:
          type: string
        pluginName:
          type: string
        version:
          type: string
        installedAt:
          type: string
          format: date-time
      required: [id, serverId, pluginId, pluginName, version, installedAt]
    InstallPluginRequest:
      type: object
      properties:
        serverId:
          type: integer
          nullable: true
        pluginId:
          type: string
      required: [pluginId]
...